+++ cookbooks/turning_technologies_turningpoint/recipes/mac.rb
@@ -18,3 +18,26 @@ dmg_package "TurningPoint" do
   package_id "com.turningtechnologies.turningpoint"
   version "5.2.1"
 end
+
+# Create directory for PowerPoint connector.
+directory "/System/Library/CFMSupport" do
+  mode 00755
+  action :create
+  recursive true
+end
+
+# Drop the PowerPoint connector library file into place.
+ cookbook_file "TurningPoint PowerPoint library" do
+  path "/System/Library/CFMSupport/TTSharedLib.1.0.0.dylib"
+  source "TTSharedLib.1.0.0.dylib"
+  mode 00755
+  :create
+end
+
+# Put TurningPoint preferences into default user profile to suppress license agreement.
+cookbook_file "TurningPoint EULA suppression" do
+  path "/System/Library/User Template/English.lproj/Library/Preferences/com.turningtechnologies.TurningPoint.plist"
+  source "com.turningtechnologies.TurningPoint.plist"
+  mode 00755
+  :create
+end