+++ cookbooks/turning_technologies_turningpoint/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file is used to list changes made in each version of turning_technologies_turningpoint.
 
+## 0.2.0:
+
+* Added PowerPoint connector library.
+* Suppressed EULA with plist file.
+
 ## 0.1.0:
 
 * Initial release of turning_technologies_turningpoint